Move logger service
